+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **All WoW Versions:**  
 - Added option to disable and hide the UI Errors Frame to /lucky config  
 - Updated deDE locales  
+- Updated frFR locales (Credits: Oniria)  
 
 **Retail:**  
 - Added option to auto accept role check popups in LFG to /lucky config