+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿## Version 4.13 [In Development]
+﻿## Version 4.13 [13-05-2026]
 
 **All WoW Versions:**  
 - Added option to disable and hide the UI Errors Frame to /lucky config  
@@ -10,7 +10,7 @@
 - Added option to enable double-clicking to signup in LFG to /lucky config  
 - Added skin for the Blizzard Cooldown Manager settings panel  
 - Removed BetterCooldownManager profile  
-- Updated SkironCooldownManager profile for all 40 specs  
+- Updated SkironCooldownManager profile for all 40 specs (Requires alpha)  
 - Updated the edit mode specialization strings for all 40 specs (Advanced)  
 
 ## Version 4.12 [06-05-2026]